refactor: 정적 데이터 도메인 분리
