+++ src/main/java/com/odiga/fiesta/companion/repository/CompanionRepository.java
@@ -1,8 +1,8 @@
-package com.odiga.fiesta.festival.repository;
+package com.odiga.fiesta.companion.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.odiga.fiesta.festival.domain.Companion;
+import com.odiga.fiesta.companion.domain.Companion;
 
 public interface CompanionRepository extends JpaRepository<Companion, Long> {
 }
