+++ src/main/java/com/odiga/fiesta/festival/dto/response/CompanionResponse.java
@@ -2,7 +2,7 @@ package com.odiga.fiesta.festival.dto.response;
 
 import static lombok.AccessLevel.*;
 
-import com.odiga.fiesta.festival.domain.Companion;
+import com.odiga.fiesta.companion.domain.Companion;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
