+++ src/main/java/com/odiga/fiesta/festival/service/CategoryService.java
@@ -5,9 +5,9 @@ import java.util.List;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.odiga.fiesta.festival.domain.Category;
+import com.odiga.fiesta.category.domain.Category;
 import com.odiga.fiesta.festival.dto.response.CategoryResponse;
-import com.odiga.fiesta.festival.repository.CategoryRepository;
+import com.odiga.fiesta.category.repository.CategoryRepository;
 
 import lombok.RequiredArgsConstructor;
