+++ src/main/java/com/odiga/fiesta/festival/service/CompanionService.java
@@ -5,9 +5,9 @@ import java.util.List;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.odiga.fiesta.festival.domain.Companion;
+import com.odiga.fiesta.companion.domain.Companion;
 import com.odiga.fiesta.festival.dto.response.CompanionResponse;
-import com.odiga.fiesta.festival.repository.CompanionRepository;
+import com.odiga.fiesta.companion.repository.CompanionRepository;
 
 import lombok.RequiredArgsConstructor;
