+++ src/main/java/com/odiga/fiesta/festival/service/PriorityService.java
@@ -5,9 +5,9 @@ import java.util.List;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.odiga.fiesta.festival.domain.Priority;
+import com.odiga.fiesta.priority.domain.Priority;
 import com.odiga.fiesta.festival.dto.response.PriorityResponse;
-import com.odiga.fiesta.festival.repository.PriorityRepository;
+import com.odiga.fiesta.priority.repository.PriorityRepository;
 
 import lombok.RequiredArgsConstructor;
