+++ src/main/java/com/odiga/fiesta/mood/repository/MoodRepository.java
@@ -1,8 +1,8 @@
-package com.odiga.fiesta.festival.repository;
+package com.odiga.fiesta.mood.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.odiga.fiesta.festival.domain.Mood;
+import com.odiga.fiesta.mood.domain.Mood;
 
 public interface MoodRepository extends JpaRepository<Mood, Long> {
 }
