+++ src/main/java/com/odiga/fiesta/priority/repository/PriorityRepository.java
@@ -1,8 +1,8 @@
-package com.odiga.fiesta.festival.repository;
+package com.odiga.fiesta.priority.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.odiga.fiesta.festival.domain.Priority;
+import com.odiga.fiesta.priority.domain.Priority;
 
 public interface PriorityRepository extends JpaRepository<Priority, Long> {
 }
