+++ src/test/java/com/odiga/fiesta/festival/service/CategoryServiceTest.java
@@ -11,9 +11,9 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 
 import com.odiga.fiesta.MockTestSupport;
-import com.odiga.fiesta.festival.domain.Category;
+import com.odiga.fiesta.category.domain.Category;
 import com.odiga.fiesta.festival.dto.response.CategoryResponse;
-import com.odiga.fiesta.festival.repository.CategoryRepository;
+import com.odiga.fiesta.category.repository.CategoryRepository;
 
 class CategoryServiceTest extends MockTestSupport {
